Handle error

+++ backend/controllers/api/apiDeleteRoutes.js
@@ -1,30 +1,39 @@
 const Project = require("../../models/projectsSchema");
 const Task = require("../../models/taskSchema");
 
-const deleteProjectApi = async (req, res) => {
+const deleteProjectApi = async (req, res, next) => {
   try {
     const { id } = req.body;
 
-    if (!id.trim()) {
-      res.status(400).json({ error: "Id is required" });
+    if (!id) {
+      throw new Error("Project id is required");
     }
     const project = await Project.findOneAndDelete({ _id: id });
+
+    if (!project) {
+      throw new Error("Invalid project id");
+    }
     const projects = await Project.find();
+
+    // delete task related to that project
     await Task.deleteMany({ projectName: project.name });
 
     res.status(204).json({ projects });
   } catch (error) {
-    res.status(400).json({ error: error.message });
+    next(error);
   }
 };
 
-const deleteTaskApi = async (req, res) => {
-  const { id } = req.body;
+const deleteTaskApi = async (req, res, next) => {
   try {
+    const { id } = req.body;
+    if (!id) {
+      throw new Error("Task Id is not present");
+    }
     const deletedTask = await Task.deleteOne({ _id: id });
     res.status(201).json({ msg: "Task deleted", task: deletedTask });
   } catch (error) {
-    res.status(500).json({ error: "It's not you. It's me" });
+    next(error);
   }
 };
